Update SinginOperationAdmin.php
edite datatype for id from
 int =>unsignedBigInteger

+++ app/Models/SinginOperationAdmin.php
@@ -14,7 +14,7 @@ class SinginOperationAdmin extends Model
     public $table="singin_operation_admins" ;
 
     protected $casts = [
-        'id' =>'integer' ,
+        'id' =>'unsignedBigInteger' ,
         'admin_id' =>'integer',
         'ip' =>'string',
         'country' =>'string',
